refactor: move content card to shared component to display content with reference hover card. used in viewpoint page display card, replycard, viewpointcard
